third_party/WebKit/LayoutTests/fast/events/pointerevents/pointer-event-mouse-coords-in-zoom-and-scroll.html

<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
    #pusher {
        width: 1000px;
        height: 1000px;
        outline: 1px solid black;
    }
</style>
<div id="console"></div>
<div id="testArea">
    <div id="pusher">This box is here to create scrollbars.</div>
</div>
<script>
    const floatPrecision = 0.00001;

    function sendMouseDownAt(x, y) {
        return new Promise(function(resolve, reject) {
            if (window.chrome && chrome.gpuBenchmarking) {
              var pointerActions =
                  [{source: "mouse",
                    actions: [
                      { name: "pointerDown", x: x, y: y },
                      { name: "pointerUp" },
                  ]}];
              chrome.gpuBenchmarking.pointerActionSequence(pointerActions, resolve);
            }
            else {
                reject();
            }
        });
    }

    function PointerDown(event) {
        eventFired = true;
        lastEvent = event;
    }

    window.addEventListener("pointerdown", PointerDown, false);

    promise_test(async () => {
        eventFired = false;
        await sendMouseDownAt(100, 100);
        assert_true(eventFired);
        assert_equals(lastEvent.clientX, 100);
        assert_equals(lastEvent.clientY, 100);
        assert_equals(lastEvent.pageX, 100);
        assert_equals(lastEvent.pageY, 100);
    }, 'Default');

    promise_test(async () => {
        eventFired = false;
        eventSender.setPageZoomFactor(1.2);
        await sendMouseDownAt(100, 100);
        assert_true(eventFired);
        assert_approx_equals(lastEvent.clientX, 83.33333, floatPrecision);
        assert_approx_equals(lastEvent.clientY, 83.33333, floatPrecision);
        assert_approx_equals(lastEvent.pageX, 83.33333, floatPrecision);
        assert_approx_equals(lastEvent.pageY, 83.33333, floatPrecision);
        eventSender.setPageZoomFactor(1.0)
    }, 'Zoomed');

    promise_test(async () => {
        eventFired = false;
        window.scrollTo(50, 50);
        await sendMouseDownAt(100, 100);
        assert_true(eventFired);
            assert_equals(lastEvent.clientX, 100);
            assert_equals(lastEvent.clientY, 100);
            assert_equals(lastEvent.pageX, 150);
            assert_equals(lastEvent.pageY, 150);
        window.scrollTo(0, 0);
    }, 'Scrolled');

    promise_test(async () => {
        eventFired = false;
        eventSender.setPageZoomFactor(1.2);
        window.scrollTo(50, 50);
        await sendMouseDownAt(100, 100);
        assert_true(eventFired);
        assert_approx_equals(lastEvent.clientX, 83.33333, floatPrecision);
        assert_approx_equals(lastEvent.clientY, 83.33333, floatPrecision);
        assert_approx_equals(lastEvent.pageX, 133.33333, floatPrecision);
        assert_approx_equals(lastEvent.pageY, 133.33333, floatPrecision);
        eventSender.setPageZoomFactor(1.0)
        window.scrollTo(0, 0);
    }, 'Zoomed and scrolled');
</script>
